requirements.txt
gate-api==6.98.0
PyYAML==6.0.1